tests/featureWriters: move shared code to parent test package
